+++ app/src/androidTest/java/com/duckduckgo/app/di/TestAppComponent.kt
@@ -73,7 +73,7 @@ import javax.inject.Singleton
         AppUsageModule::class,
         FileModule::class,
         UncaughtExceptionModule::class,
-        PlayStoreReferralModule::class,
+        StoreReferralModule::class,
         CoroutinesModule::class,
         CertificateTrustedStoreModule::class,
         WelcomePageModule::class
