+++ app/src/main/java/com/duckduckgo/app/di/AppComponent.kt
@@ -71,7 +71,7 @@ import javax.inject.Singleton
         AppUsageModule::class,
         FileModule::class,
         UncaughtExceptionModule::class,
-        PlayStoreReferralModule::class,
+        StoreReferralModule::class,
         CoroutinesModule::class,
         CertificateTrustedStoreModule::class,
         WelcomePageModule::class
